Add tab completion for file paths.
Internally, lsdir() on the server side will be called. The usage of the
new completion feature is restrited on the function arguments which
could be used for filenames. The list could be found in keykit_language.py.
Relative and absolute paths are allowed.

Example: readfm("music/<Tab><Tab> should list your midi files in music.
Note that chdir() could be used to switch the root directory for
relative paths.